database.js: Add support for array values in where conditions
This commit adds support for array values in the `where` conditions of the `database.js` file. Previously, only single values were supported, but now, if the `value` is an array, the condition will use the `= ANY` operator instead of `=`. This change allows for more flexible querying options.

test/sql.js: Add unit test for array values in database.select

This commit adds a new unit test to the `sql.js` file. The test checks that the `database.select` method properly handles array values in the `where` conditions. It verifies that the method correctly returns the expected results when an array of values is provided as the condition for a specific property.

Note: The changes to imports and requires have been ignored in this commit message.